refactor: alert src
